+++ src/deluge/gui/views/view.cpp
@@ -59,6 +59,7 @@
 #include "model/clip/instrument_clip_minder.h"
 #include "model/consequence/consequence.h"
 #include "model/drum/drum.h"
+#include "model/instrument/cv_instrument.h"
 #include "model/instrument/instrument.h"
 #include "model/instrument/kit.h"
 #include "model/instrument/melodic_instrument.h"
@@ -1858,7 +1859,7 @@ void View::displayOutputName(Output* output, bool doBlink, Clip* clip) {
 			// No break
 
 		case OutputType::CV:
-			channel = ((NonAudioInstrument*)instrument)->channel;
+			channel = ((NonAudioInstrument*)instrument)->getChannel();
 			break;
 		}
 	}
@@ -2053,13 +2054,23 @@ yesAlignRight:
 	}
 	else if (outputType == OutputType::CV) {
 		if (display->haveOLED()) {
-			intToString(channel + 1, buffer);
+			if (channel < both) {
+				intToString(channel + 1, buffer);
+			}
+			else {
+				sprintf(buffer, "1 and 2");
+			}
 oledOutputBuffer:
 			nameToDraw = buffer;
 			goto oledDrawString;
 		}
 		else {
-			display->setTextAsNumber(channel + 1, 255, doBlink);
+			if (channel < both) {
+				display->setTextAsNumber(channel + 1, 255, doBlink);
+			}
+			else {
+				display->setText("Both");
+			}
 		}
 	}
 }
@@ -2135,7 +2146,7 @@ void View::navigateThroughPresetsForInstrumentClip(int32_t offset, ModelStackWit
 	if (outputType == OutputType::MIDI_OUT || outputType == OutputType::CV) {
 
 		NonAudioInstrument* oldNonAudioInstrument = (NonAudioInstrument*)oldInstrument;
-		int32_t newChannel = oldNonAudioInstrument->channel;
+		int32_t newChannel = oldNonAudioInstrument->getChannel();
 		int32_t newChannelSuffix;
 		if (outputType == OutputType::MIDI_OUT) {
 			newChannelSuffix = ((MIDIInstrument*)oldNonAudioInstrument)->channelSuffix;
@@ -2146,9 +2157,9 @@ void View::navigateThroughPresetsForInstrumentClip(int32_t offset, ModelStackWit
 		// CV
 		if (outputType == OutputType::CV) {
 			while (true) {
-				newChannel = (newChannel + offset) & (NUM_CV_CHANNELS - 1);
+				newChannel = CVInstrument::navigateChannels(newChannel, offset);
 
-				if (newChannel == oldNonAudioInstrument->channel) {
+				if (newChannel == oldNonAudioInstrument->getChannel()) {
 					display->displayPopup(deluge::l10n::get(deluge::l10n::String::STRING_FOR_NO_UNUSED_CHANNELS));
 					return;
 				}
@@ -2175,7 +2186,7 @@ void View::navigateThroughPresetsForInstrumentClip(int32_t offset, ModelStackWit
 			int32_t oldChannel = newChannel;
 
 			if (oldInstrumentCanBeReplaced) {
-				oldNonAudioInstrument->channel = -1; // Get it out of the way
+				oldNonAudioInstrument->setChannel(-1); // Get it out of the way
 			}
 
 			while (true) {
@@ -2213,7 +2224,7 @@ void View::navigateThroughPresetsForInstrumentClip(int32_t offset, ModelStackWit
 
 				if (newChannel == oldChannel
 				    && newChannelSuffix == ((MIDIInstrument*)oldNonAudioInstrument)->channelSuffix) {
-					oldNonAudioInstrument->channel = oldChannel; // Put it back
+					oldNonAudioInstrument->setChannel(oldChannel); // Put it back
 					display->displayPopup(deluge::l10n::get(deluge::l10n::String::STRING_FOR_NO_UNUSED_CHANNELS));
 					return;
 				}
@@ -2235,7 +2246,7 @@ void View::navigateThroughPresetsForInstrumentClip(int32_t offset, ModelStackWit
 				}
 			}
 
-			oldNonAudioInstrument->channel = oldChannel; // Put it back
+			oldNonAudioInstrument->setChannel(oldChannel); // Put it back
 		}
 
 		newInstrument =
@@ -2256,7 +2267,7 @@ void View::navigateThroughPresetsForInstrumentClip(int32_t offset, ModelStackWit
 
 			// Because these are just MIDI / CV instruments and we're changing them for all Clips, we can just change
 			// the existing Instrument object!
-			oldNonAudioInstrument->channel = newChannel;
+			oldNonAudioInstrument->setChannel(newChannel);
 			if (outputType == OutputType::MIDI_OUT) {
 				((MIDIInstrument*)oldNonAudioInstrument)->channelSuffix = newChannelSuffix;
 			}