+++ src/deluge/model/song/song.cpp
@@ -505,7 +505,7 @@ void Song::transposeAllScaleModeClips(int32_t offset, bool chromatic) {
 				    modelStack->addTimelineCounter(instrumentClip);
 				if (instrumentClip->isScaleModeClip()) {
 					if (clip->output->type == OutputType::MIDI_OUT
-					    && ((NonAudioInstrument*)clip->output)->channel == MIDI_CHANNEL_TRANSPOSE) {
+					    && ((NonAudioInstrument*)clip->output)->getChannel() == MIDI_CHANNEL_TRANSPOSE) {
 						// Must not transpose MIDI clips that are routed to transpose, ie note rows
 						// stay exactly the same.
 						// Just have to scroll the clip so that the change in song root note
@@ -645,7 +645,7 @@ void Song::replaceMusicalMode(const ScaleChange& changes, bool affectMIDITranspo
 
 	for (InstrumentClip* instrumentClip : InstrumentClips::everywhere(this)) {
 		if (!affectMIDITranspose && instrumentClip->output->type == OutputType::MIDI_OUT
-		    && ((NonAudioInstrument*)instrumentClip->output)->channel == MIDI_CHANNEL_TRANSPOSE) {
+		    && ((NonAudioInstrument*)instrumentClip->output)->getChannel() == MIDI_CHANNEL_TRANSPOSE) {
 			// Must not transpose MIDI clips that are routed to transpose.
 			continue;
 		}
@@ -1319,10 +1319,9 @@ Error Song::readFromFile(Deserializer& reader) {
 
 	uint64_t newTimePerTimerTick = (uint64_t)1 << 32; // TODO: make better!
 
-	// reverb mode
-	if (song_firmware_version < FirmwareVersion::official({4, 1, 4})) {
-		AudioEngine::reverb.setModel(deluge::dsp::Reverb::Model::FREEVERB);
-	}
+	// reverb mode is freeverb for songs that predate having multiple options. New songs will set it to mutable anyway
+	// so this is only used as a fallback
+	AudioEngine::reverb.setModel(deluge::dsp::Reverb::Model::FREEVERB);
 
 	if (!reader.match('{'))
 		return Error::FILE_CORRUPTED;
@@ -3063,7 +3062,7 @@ int32_t Song::getMaxMIDIChannelSuffix(int32_t channel) {
 	for (Output* output = firstOutput; output; output = output->next) {
 		if (output->type == OutputType::MIDI_OUT) {
 			MIDIInstrument* instrument = (MIDIInstrument*)output;
-			if (instrument->channel == channel) {
+			if (instrument->getChannel() == channel) {
 				int32_t suffix = instrument->channelSuffix;
 				if (suffix < -1 || suffix >= 26) {
 					continue; // Just for safety
@@ -3523,47 +3522,18 @@ Instrument* Song::getInstrumentFromPresetSlot(OutputType outputType, int32_t cha
 
 	if (searchNonHibernating) {
 		for (Output* thisOutput = firstOutput; thisOutput; thisOutput = thisOutput->next) {
-			if (thisOutput->type == outputType) {
-
-				bool match;
-
-				if (outputType == OutputType::SYNTH || outputType == OutputType::KIT) {
-					match = !strcasecmp(name, thisOutput->name.get())
-					        && !strcasecmp(dirPath, ((Instrument*)thisOutput)->dirPath.get());
-				}
-
-				else {
-					match = (((NonAudioInstrument*)thisOutput)->channel == channel
-					         && (outputType == OutputType::CV
-					             || ((MIDIInstrument*)thisOutput)->channelSuffix == channelSuffix));
-				}
-
-				if (match) {
-					return (Instrument*)thisOutput;
-				}
+			bool match = thisOutput->matchesPreset(outputType, channel, channelSuffix, name, dirPath);
+			if (match) {
+				return (Instrument*)thisOutput;
 			}
 		}
 	}
 
 	if (searchHibernating) {
 		for (Output* thisOutput = firstHibernatingInstrument; thisOutput; thisOutput = thisOutput->next) {
-			if (thisOutput->type == outputType) {
-
-				bool match;
-
-				if (outputType == OutputType::SYNTH || outputType == OutputType::KIT) {
-					match = !strcasecmp(name, thisOutput->name.get())
-					        && !strcasecmp(dirPath, ((Instrument*)thisOutput)->dirPath.get());
-				}
-
-				else {
-					match = (((NonAudioInstrument*)thisOutput)->channel == channel
-					         && ((MIDIInstrument*)thisOutput)->channelSuffix == channelSuffix);
-				}
-
-				if (match) {
-					return (Instrument*)thisOutput;
-				}
+			bool match = thisOutput->matchesPreset(outputType, channel, channelSuffix, name, dirPath);
+			if (match) {
+				return (Instrument*)thisOutput;
 			}
 		}
 	}
@@ -3920,7 +3890,7 @@ bool Song::doesNonAudioSlotHaveActiveClipInSession(OutputType outputType, int32_
 
 			Instrument* instrument = (Instrument*)clip->output;
 
-			if (instrument->type == outputType && ((NonAudioInstrument*)instrument)->channel == slot
+			if (instrument->type == outputType && ((NonAudioInstrument*)instrument)->getChannel() == slot
 			    && (outputType == OutputType::CV || ((MIDIInstrument*)instrument)->channelSuffix == subSlot)) {
 				return true;
 			}
@@ -4305,7 +4275,7 @@ MIDIInstrument* Song::grabHibernatingMIDIInstrument(int32_t channel, int32_t cha
 		toReturn->setActiveClip(nullptr, PgmChangeSend::NEVER); // Not really necessary?
 		toReturn->inValidState = false;
 
-		toReturn->channel = channel;
+		toReturn->setChannel(channel);
 		toReturn->channelSuffix = channelSuffix;
 	}
 	return toReturn;
@@ -4635,8 +4605,8 @@ Output* Song::navigateThroughPresetsForInstrument(Output* output, int32_t offset
 
 		NonAudioInstrument* oldNonAudioInstrument = (NonAudioInstrument*)oldInstrument;
 
-		int32_t oldChannel = oldNonAudioInstrument->channel;
-		int32_t newChannel = oldNonAudioInstrument->channel;
+		int32_t oldChannel = oldNonAudioInstrument->getChannel();
+		int32_t newChannel = oldNonAudioInstrument->getChannel();
 
 		int32_t oldChannelSuffix, newChannelSuffix;
 		if (outputType == OutputType::MIDI_OUT) {
@@ -4647,7 +4617,7 @@ Output* Song::navigateThroughPresetsForInstrument(Output* output, int32_t offset
 		// CV
 		if (outputType == OutputType::CV) {
 			do {
-				newChannel = (newChannel + offset) & (NUM_CV_CHANNELS - 1);
+				newChannel = CVInstrument::navigateChannels(newChannel, offset);
 
 				if (newChannel == oldChannel) {
 cantDoIt:
@@ -4661,7 +4631,7 @@ cantDoIt:
 		// Or MIDI
 		else {
 
-			oldNonAudioInstrument->channel = -1; // Get it out of the way
+			oldNonAudioInstrument->setChannel(-1); // Get it out of the way
 
 			do {
 				newChannelSuffix += offset;
@@ -4683,14 +4653,14 @@ cantDoIt:
 				}
 
 				if (newChannel == oldChannel && newChannelSuffix == oldChannelSuffix) {
-					oldNonAudioInstrument->channel = oldChannel; // Put it back
+					oldNonAudioInstrument->setChannel(oldChannel); // Put it back
 					goto cantDoIt;
 				}
 
 			} while (
 			    currentSong->getInstrumentFromPresetSlot(outputType, newChannel, newChannelSuffix, NULL, NULL, false));
 
-			oldNonAudioInstrument->channel = oldChannel; // Put it back, before switching notes off etc
+			oldNonAudioInstrument->setChannel(oldChannel); // Put it back, before switching notes off etc
 		}
 
 		if (oldNonAudioInstrument->getActiveClip() && playbackHandler.isEitherClockActive()) {
@@ -4699,7 +4669,7 @@ cantDoIt:
 
 		// Because these are just MIDI / CV instruments and we're changing them for all Clips, we can just change
 		// the existing Instrument object!
-		oldNonAudioInstrument->channel = newChannel;
+		oldNonAudioInstrument->setChannel(newChannel);
 		if (outputType == OutputType::MIDI_OUT) {
 			((MIDIInstrument*)oldNonAudioInstrument)->channelSuffix = newChannelSuffix;
 		}
@@ -4811,7 +4781,7 @@ Instrument* Song::changeOutputType(Instrument* oldInstrument, OutputType newOutp
 	// MIDI / CV
 	if (newOutputType == OutputType::MIDI_OUT || newOutputType == OutputType::CV) {
 
-		int32_t numChannels = (newOutputType == OutputType::MIDI_OUT) ? 16 : NUM_CV_CHANNELS;
+		int32_t numChannels = (newOutputType == OutputType::MIDI_OUT) ? 16 : kNumCVInstrumentChannels;
 
 		while (true) {
 
@@ -4819,7 +4789,7 @@ Instrument* Song::changeOutputType(Instrument* oldInstrument, OutputType newOutp
 				break;
 			}
 
-			newSlot = (newSlot + 1) & (numChannels - 1);
+			newSlot = (newSlot + 1) % numChannels;
 			newSubSlot = -1;
 
 			// If we've searched all channels...
@@ -5099,7 +5069,7 @@ void Song::getNoteLengthName(StringBuf& buffer, uint32_t noteLength, char const*
 Instrument* Song::getNonAudioInstrumentToSwitchTo(OutputType newOutputType, Availability availabilityRequirement,
                                                   int16_t newSlot, int8_t newSubSlot,
                                                   bool* instrumentWasAlreadyInSong) {
-	int32_t numChannels = (newOutputType == OutputType::MIDI_OUT) ? 16 : NUM_CV_CHANNELS;
+	int32_t numChannels = (newOutputType == OutputType::MIDI_OUT) ? 16 : kNumCVInstrumentChannels;
 	Instrument* newInstrument;
 	int32_t oldSlot = newSlot;